% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUB.datasets.R
\name{HUB.datasets}
\alias{HUB.datasets}
\title{Show the datasets available on the IPCC hub}
\usage{
HUB.datasets(pattern = "", full.info = FALSE)
}
\arguments{
\item{pattern}{Optional. Pattern in the dataset name as passed to function \code{\link{grep}} (case-insensitive).}

\item{full.info}{Default to FALSE. If TRUE, the complete URL is also displayed.}
}
\value{
The datasets table, in the form of a \code{data.frame}
}
\description{
Lists the harmonized datasets that are accessible from the IPCC hub.
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in datasets
# str(HUB.datasets())
# (sets <- HUB.datasets())
# names(sets)
# sets$OBSERVATIONS
# # Using argument pattern
# HUB.datasets(pattern = "E-OBS.*v21")
# # Showing the full url of the datasets
# HUB.datasets(pattern = "E-OBS", full.info = TRUE)
}
\author{
J Baño-Medina
}
