% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHubData.R
\name{loadHubData}
\alias{loadHubData}
\title{Show the datasets available on the IPCC hub}
\usage{
loadHubData(label)
}
\arguments{
\item{label}{Optional. Pattern in the dataset name as passed to function \code{\link{grep}} (case-insensitive).}
}
\value{
The datasets table, in the form of a \code{data.frame}
}
\description{
Lists the harmonized datasets that are accessible from the IPCC hub.
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Example with E-OBS v21
# library(loadeR)
# url <- loadHubData("E-OBS")
# data <- loadGridData(dataset = url, years = 2000)
}
\author{
J Baño-Medina
}
